Move replacement code to Replace folder and refactored code
