+++ Models/Stock/Genotype.cs
@@ -1,6 +1,7 @@
 namespace Models.GrazPlan
 {
     using Models.Core;
+    using Models.Core.Replace;
     using Models.Core.Run;
     using StdUnits;
     using System;
